previous_work/core/requirements_vision.txt
absl-py==2.1.0
actionlib==1.14.0
angles==1.9.13
apriltag==0.0.16
astunparse==1.6.3
base_local_planner==1.17.3
beautifulsoup4==4.12.3
blinker==1.8.2
bondpy==1.8.6
cachetools==5.4.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
                deepface==0.0.92
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.9.0
docutils==0.20.1
drone-wrapper==1.4.2
dynamic-reconfigure==1.7.3
dynamixel_sdk==3.7.51
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
futures==3.0.5
gast==0.4.0
gazebo_model_attachment_plugin==1.0.2
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gazebo_video_monitor_utils==0.7.1
gdown==5.2.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.33.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.4
gunicorn==23.0.0
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
image-geometry==1.16.2
importlib_metadata==8.2.0
importlib_resources==6.4.0
interactive-markers==1.12.0
itsdangerous==2.2.0
Jinja2==3.1.4
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==2.13.1
kiwisolver==1.4.5
laser_geometry==1.6.7
libclang==18.1.1
llvmlite==0.41.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
mavros==1.19.0
message-filters==1.16.0
miro2_msg==0.1.0
mpmath==1.3.0
mtcnn==0.1.1
networkx==3.1
                numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
                pandas==2.0.3
pillow==10.4.0
pkg_resources==0.0.0
pr2_power_board==1.1.10
protobuf==3.20.3
py-trees==0.7.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.26.1
PyGObject==3.48.2
pyparsing==3.1.2
pyspin==1.1.1
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytz==2024.1
PyYAML==6.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
resource_retriever==1.12.7
retina-face==0.0.17
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.1
rospy==1.16.0
rosserial_arduino==0.9.2
rosserial_client==0.9.2
rosserial_python==0.9.2
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-console==0.4.12
rqt-controller-manager==0.20.0
rqt-drone-teleop==1.4.2
rqt-ez-publisher==0.6.1
rqt-ground-robot-teleop==1.4.2
rqt-image-view==0.4.17
rqt-joint-trajectory-controller==0.22.0
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-multiplot==0.0.12
rqt-pr2-dashboard==0.4.1
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_py_trees==0.4.1
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.25
safetensors==0.4.4
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
soupsieve==2.5
sympy==1.13.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-gpu==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tokenizers==0.19.1
topic-tools==1.16.0
            torch==2.4.0
tqdm==4.66.5
            transformers==4.44.0
triton==3.0.0
turtlebot3_example==1.2.5
turtlebot3_teleop==1.2.5
typing_extensions==4.5.0
tzdata==2024.1
unique_id==1.0.6
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
xacro==1.14.18
zipp==3.20.0
        mediapipe
        defisheye